+++ src/jcpsantiago/arqivist/api/confluence/handlers.clj
@@ -6,10 +6,15 @@
   * lifecycle endpoints used for the 'installed', 'enabled' and 'uninstalled' events
   "
   (:require
+   [clojure.spec.alpha :as spec]
    [com.brunobonacci.mulog :as mulog]
    [org.httpkit.client :as httpkit]
+   [jcpsantiago.arqivist.api.confluence.db :as db]
    [jcpsantiago.arqivist.api.confluence.utils :as utils]
-   [next.jdbc.sql :as sql]))
+   [jcpsantiago.arqivist.api.slack.specs :as slack-specs]
+   [jsonista.core :as jsonista]
+   [next.jdbc.sql :as sql]
+   [ring.util.response :refer [response content-type]]))
 
 (defn app-descriptor-json
   "
@@ -24,8 +29,7 @@
   (fn [_]
     (let [env (get-in system [:atlassian-env])]
       (mulog/log ::serving-descriptor-json :local-time (java.time.LocalDateTime/now))
-      {:status 200
-       :body
+      (response
        {:key (:descriptor-key env)
         :name "The Arqivist - Slack conversations become Confluence pages"
         :description "Create Confluence pages from Slack conversations."
@@ -50,7 +54,7 @@
            :key "theArqivistMetadata"
            ;; this one must be snake_case... this is not documented, I just tried and failed a few times
            :keyConfigurations [{:propertyKey "the_arqivist_props"
-                                :extractions (mapv utils/content-properties-extraction (utils/content-properties-ks))}]}]}}})))
+                                :extractions (mapv utils/content-properties-extraction (utils/content-properties-ks))}]}]}}))))
 
 (defn installed
   "
@@ -60,8 +64,9 @@
 
   It creates a new entry in the atlassian_tenants table with the keys needed for the app to work.
   "
-  [lifecycle-payload tenant_id system]
+  [lifecycle-payload atlassian_tenant system]
   (let [db-connection (:db-connection system)
+        tenant_id (:atlassian_tenants/id atlassian_tenant)
         {:keys [key clientKey accountId sharedSecret baseUrl displayUrl productType
                 description serviceEntitlementNumber oauthClientId]} lifecycle-payload
         url-short (utils/base-url-short baseUrl)
@@ -78,34 +83,72 @@
                         :service_entitlement_number serviceEntitlementNumber
                         :oauth_client_id oauthClientId}]
 
-    (mulog/log ::atlassian-installed :base-url baseUrl :local-time (java.time.LocalDateTime/now))
-
     (if (nil? tenant_id)
+      ;; tenant did not connect before, we'll insert it into the db and create a space
+      (let [inserted (db/insert-new-atlassian-tenant! data-to-insert db-connection)]
+        (if inserted
+          (response "OK")
+          {:status 500 :body "Couldn't insert credentials into db."}))
 
-      (do
-        (mulog/log ::inserting-new-atlassian-tenant :base-url baseUrl)
-        (sql/insert! db-connection :atlassian_tenants data-to-insert {:return-keys true})
-        {:status 200 :body "OK"})
-
-      (do
-        (mulog/log ::updating-existing-atlassian-tenant :base-url baseUrl)
-        (sql/update! db-connection :atlassian_tenants data-to-insert {:id tenant_id})
-        {:status 200 :body "OK"}))))
+      ;; else branch:
+      ;; tenant already exists in the db, we'll update our data instead
+      ;; this may happen because the tenant has new credentials from Atlassian
+      (let [updated (db/update-existing-atlassian-tenant! data-to-insert tenant_id db-connection)]
+        (if updated
+          (response "OK")
+          {:status 500 :body "Couldn't update tenant's credentials."})))))
 
 (defn enabled
   "
   Ring handler for the 'enabled' event issued after 'installed' is successful.
   It creates a space in the user's Confluence instance, used later to store all archived conversations.
+  We can only create spaces, or interact with Confluence at all after Atlassian issues this event.
   "
-  [lifecycle-payload _ system]
-  (mulog/log ::atlassian-enabled :base-url (:baseUrl lifecycle-payload) :local-time (java.time.LocalDateTime/now))
+  [lifecycle-payload atlassian_tenant system]
+  (let [{:keys [db-connection atlassian-env]} system
+        {:keys [descriptor-key space-key]} atlassian-env
+        shared_secret (:atlassian_tenants/shared_secret atlassian_tenant)
+        tenant_id (:atlassian_tenants/id atlassian_tenant)]
+    (try
+      (if (utils/space-exists? atlassian-env lifecycle-payload shared_secret)
+        (do
+          (mulog/log ::skipped-creating-arqivist-confluence-space
+                     :local-time (java.time.LocalDateTime/now))
+          (response "OK"))
 
-  (let [res (utils/create-space! (get-in system [:env :atlassian :descriptor-key]) lifecycle-payload)]
+        (let [;; We create a space for storing conversations, but never delete it even during uninstall,
+              ;; this way users keep their messages no matter what
+              create-space-res (utils/create-space! descriptor-key lifecycle-payload shared_secret space-key)]
 
-    (if (= (:status res) 200)
-      {:status 200 :body "OK"}
-      ;; TODO: something more useful here? This endpint is not consumed by anyone though, only the Atlassian bots
-      {:status 500 :body "ERROR"})))
+          (if (= (:status create-space-res) 200)
+            (do
+              (mulog/log ::created-arqivist-confluence-space
+                         :success :true
+                         :local-time (java.time.LocalDateTime/now))
+              (response "OK"))
+
+            ;; in case we can't create a space, it's time to clean-up
+            (do
+              (mulog/log ::created-arqivist-confluence-space
+                         :success :false
+                         :confluence-http-status (:status create-space-res)
+                         :error (:body create-space-res)
+                         :descriptor-key descriptor-key
+                         :local-time (java.time.LocalDateTime/now))
+
+              ;; clean-up after the failure, we don't want to keep unnecessary credentials in our db
+              (sql/delete! db-connection :atlassian_tenants {:id tenant_id})
+
+              (mulog/log ::cleaning-up-inserting-new-tenant
+                         :success :true
+                         :local-time (java.time.LocalDateTime/now))
+              {:status 500 :body "Couldn't create Confluence space."}))))
+
+      (catch Exception e
+        (mulog/log ::enabled-app
+                   :success :false
+                   :error e)
+        {:status 500 :body "Couldn't create Confluence space."}))))
 
 (defn uninstalled
   "
@@ -113,35 +156,72 @@
     * deletes all rows in db related to the tenant originating the request
     * uninstalls the app from the Slack workspace connected to the tenant
   "
-  [lifecycle-payload tenant_id system]
+  [_ atlassian_tenant system]
   (let [db-connection (:db-connection system)
-        {:keys [baseUrl]} lifecycle-payload
+        tenant_id (:atlassian_tenants/id atlassian_tenant)
         ;; Slack credentials linked to the tenant, used below to uninstall
         {:keys [:slack_teams/access_token
                 :slack_teams/team_name
-                :slack_teams/external_team_id]}
+                :slack_teams/external_team_id] :as slack_team}
         (-> (sql/find-by-keys db-connection :slack_teams {:atlassian_tenant_id tenant_id})
             first)]
 
     ;; drop the row corresponding to the current tenant, plus the row in  'slack_teams'
-    ;; TODO: wrap in try/catch
-    (mulog/log ::uninstalling-app
-               :base-url baseUrl
-               :slack-team-id external_team_id
-               :slack-team-name team_name
-               :local-time (java.time.LocalDateTime/now))
+    (mulog/with-context
+     {:slack-team-id external_team_id
+      :slack-team-name team_name}
 
-    (sql/delete! db-connection :atlassian_tenants {:id tenant_id})
+     (try
+       (sql/delete! db-connection :atlassian_tenants {:id tenant_id})
+       (mulog/log ::atlassian-tenant-dropped-from-db
+                  :success :true
+                  :local-time (java.time.LocalDateTime/now))
 
-    ;; TODO: wrap in try/catch
-    @(httpkit/get
-      (str "https://slack.com/api/apps.uninstall?"
-           "client_id=" (get-in system [:env :slack :arqivist-slack-client-id])
-           "&client_secret=" (get-in system [:env :slack :arqivist-slack-client-secret])
-           {:headers {"Content-Type" "application/json; charset=utf-8"}
-            :oauth-token access_token}))
+       (if (empty? slack_team)
+         ;; it's possible the user never connected their Slack workspace at all,
+         ;; in which case we don't need to do anything else
+         (response "OK")
 
-    {:status 200 :body "OK"}))
+         ;; if there's a slack team connected, we have to uninstall the app
+         (let [res (-> @(httpkit/get
+                         (str "https://slack.com/api/apps.uninstall?"
+                              "client_id=" (get-in system [:env :slack :arqivist-slack-client-id])
+                              "&client_secret=" (get-in system [:env :slack :arqivist-slack-client-secret]))
+                         {:headers {"Content-Type" "application/json; charset=utf-8"}
+                          :oauth-token access_token})
+                       :body
+                       (jsonista/read-value jsonista/keyword-keys-object-mapper))]
+           (cond
+             (spec/invalid? (spec/conform ::slack-specs/apps-uninstall res))
+             (do
+               (mulog/log ::uninstalled-from-slack
+                          :success :false
+                          :error "Slack API response did not conform to spec"
+                          :spec-explain (spec/explain ::slack-specs/apps-uninstall res)
+                          :local-time (java.time.LocalDateTime/now))
+               (-> {:status 500 :body "Couldn't uninstall from Slack, got invalid response"} (content-type "text-plain")))
+
+             (:ok res)
+             (do
+               (mulog/log ::uninstalled-from-slack
+                          :success :true
+                          :local-time (java.time.LocalDateTime/now))
+               (-> "OK" response (content-type "text/plain")))
+
+             :else
+             (do
+               (mulog/log ::uninstalled-from-slack
+                          :success :false
+                          :error (:error res)
+                          :local-time (java.time.LocalDateTime/now))
+               (-> {:status 500 :body "Couldn't uninstall from Slack!"} (content-type "text-plain"))))))
+
+       (catch Exception e
+         ;; TODO: send a Slack message to the admin user informing them this failed, and they need to manually remove the app
+         (mulog/log ::uninstalling-app
+                    :success :false
+                    :error (.getMessage e)
+                    :local-time (java.time.LocalDateTime/now)))))))
 
 (defn lifecycle
   "
@@ -152,15 +232,23 @@
     (let [db-connection (:db-connection system)
           lifecycle-payload (:body-params request)
           event-type (:eventType lifecycle-payload)
-          {:keys [:atlassian_tenants/tenant_id]} (-> (sql/find-by-keys
-                                                      db-connection
-                                                      :atlassian_tenants
-                                                      {:base_url_short (utils/base-url-short (:baseUrl lifecycle-payload))}
-                                                      {:columns [[:id :tenant_id]]})
-                                                     first)]
-      (mulog/log ::lifecycle-event :event-type event-type :base-url (:baseUrl lifecycle-payload) :local-time (java.time.LocalDateTime/now))
+          base-url (:baseUrl lifecycle-payload)
 
-      (case event-type
-        "installed" (installed lifecycle-payload tenant_id system)
-        "enabled" (enabled lifecycle-payload tenant_id system)
-        "uninstalled" (uninstalled lifecycle-payload tenant_id system)))))
+          {:keys [:atlassian_tenants/tenant_id] :as atlassian_tenant}
+          (-> (sql/find-by-keys
+               db-connection
+               :atlassian_tenants
+               {:base_url_short (utils/base-url-short (:baseUrl lifecycle-payload))})
+              first)]
+
+      (mulog/log ::lifecycle-event :event-type event-type :base-url base-url :local-time (java.time.LocalDateTime/now))
+
+      (mulog/with-context
+       {:base-url (:baseUrl lifecycle-payload)
+        :event-type event-type
+        :tenant-id tenant_id}
+
+       (case event-type
+         "installed" (installed lifecycle-payload atlassian_tenant system)
+         "enabled" (enabled lifecycle-payload atlassian_tenant system)
+         "uninstalled" (uninstalled lifecycle-payload atlassian_tenant system))))))