+++ src/jcpsantiago/arqivist/api/confluence/router.clj
@@ -1,6 +1,7 @@
 (ns jcpsantiago.arqivist.api.confluence.router
   "Reitit routes for interacting with Confluence."
   (:require
+   [jcpsantiago.arqivist.middleware :as middleware-arqivist]
    [jcpsantiago.arqivist.api.confluence.handlers :as handlers]
    [jcpsantiago.arqivist.api.confluence.specs :as specs]))
 
@@ -30,7 +31,8 @@
    ;; FIXME: These routes need different kind of validation than the non-lifecycle routes
    ;; https://developer.atlassian.com/cloud/confluence/security-for-connect-apps/#validating-installation-lifecycle-requests
    ["/installed"
-    {:post
+    {:middleware [[middleware-arqivist/verify-atlassian-lifecycle :verify-atlassian-jwt]]
+     :post
      {:summary "Webhook for the 'install' event"
       :description "Webhook for the `install` event<br>
                     Triggered in various events, but the most important is the initial app installation to a site.<br>
@@ -48,7 +50,8 @@
       :handler (handlers/lifecycle system)}}]
 
    ["/uninstalled"
-    {:post
+    {:middleware [[middleware-arqivist/verify-atlassian-lifecycle :verify-atlassian-jwt]]
+     :post
      {:summary "Webhook for the 'uninstall' event"
       :description "Webhook for the `uninstall` event when a user uninstalls the app from a site."
       :parameters {:body ::specs/lifecycle}