+++ src/jcpsantiago/arqivist/api/confluence/utils.clj
@@ -7,7 +7,6 @@
    [buddy.core.hash :refer [sha256]]
    [buddy.sign.jwt :as jwt]
    [clojure.string :as string]
-   [com.brunobonacci.mulog :as mulog]
    [org.httpkit.client :as httpkit]
    [java-time :as t]
    [jsonista.core :as jsonista]))
@@ -50,9 +49,47 @@
      {"Content-Type" "application/json; charset=utf-8"
       ;; TODO: review why we need this, and document it.
       ;; I used this in the previous iteration after a lot of trial and error
-      "X-Atlassian-Token" "no-check"
+      ;; "X-Atlassian-Token" "no-check"
       "Authorization" (str "JWT " jwt-token)}}
     opts)))
+
+(defn get-spaces!
+  "
+  Calls the Confluence API and retrieves a list of existing spaces.
+  See the example response payload in /test/jcpsantiago/get_confluence_spaces.json.
+  "
+  [descriptor-key lifecycle-payload shared_secret]
+  (let [canonical-url "/rest/api/space"
+        base-url (:baseUrl lifecycle-payload)
+        jwt-token (atlassian-jwt descriptor-key shared_secret "GET" canonical-url)]
+    (httpkit/get
+     (str base-url canonical-url)
+     (opts-with-jwt jwt-token))))
+
+(defn space-exists?
+  "
+  Checks if a 'space-key' exists in a Confluence instance.
+  Returns true/false, and throws if get-spaces! has non-200 status.
+  "
+  [atlassian-env lifecycle-payload shared_secret]
+  (let [{:keys [descriptor-key space-key]} atlassian-env
+        space-res @(get-spaces! descriptor-key lifecycle-payload shared_secret)]
+    (if (= (:status space-res) 200)
+      (if-let [space-results (-> space-res
+                                 :body
+                                 (jsonista/read-value jsonista/keyword-keys-object-mapper)
+                                 :results)]
+
+        (if (empty? (filter #(= (:key %) space-key) space-results))
+          false
+          true)
+
+        ;; in the extremely unlikely event the :results are completely empty 
+        ;; i.e. the instance has no spaces, we will assume that's real and our space doesn't exist yet
+        false)
+
+      ;; In case the call fails (non 200 status), for some reason. Really edge-casey here.
+      (throw (Exception. "Couldn't get spaces from Confluence. Body: " (:body space-res))))))
 
 (defn create-space!
   "
@@ -62,28 +99,21 @@
   Takes a descriptor key provided by Atlassian for The Arqivist and a lifecycle payload as described
   in https://developer.atlassian.com/cloud/confluence/connect-app-descriptor/#lifecycle
   "
-  [descriptor-key lifecycle-payload]
+  [descriptor-key lifecycle-payload shared_secret space-key]
   (let [canonical-url "/rest/api/space"
         base-url (:baseUrl lifecycle-payload)
-        jwt-token (atlassian-jwt descriptor-key (:sharedSecret lifecycle-payload) "POST" canonical-url)
-        res @(httpkit/post
-              (str base-url canonical-url)
-              (opts-with-jwt jwt-token {:body (jsonista/write-value-as-string
-                                               ;; NOTE: for now we're hardcoding this, not sure if users will care about
-                                               ;; using their own names for this space
-                                               {:key "SLCKARQVST"
-                                                :name "Archived Slack threads"
-                                                :description
-                                                {:plain
-                                                 {:value "Space for Slack conversations archived by The Arqivist bot"
-                                                  :representation "plain"}}})}))]
-    (if (= (:status res) 200)
-      (do
-        (mulog/log ::confluence-space-created :ok :true :base-url base-url)
-        {:ok true})
-      (do
-        (mulog/log ::confluence-space-created :ok :false :base-url base-url :exception (:body res))
-        {:ok false}))))
+        jwt-token (atlassian-jwt descriptor-key shared_secret "POST" canonical-url)]
+    (httpkit/post
+     (str base-url canonical-url)
+     (opts-with-jwt jwt-token {:body (jsonista/write-value-as-string
+                                      ;; NOTE: for now we're hardcoding this, not sure if users will care about
+                                      ;; using their own names for this space
+                                      {:key space-key
+                                       :name "Archived Slack threads"
+                                       :description
+                                       {:plain
+                                        {:value "Space for Slack conversations archived by The Arqivist bot"
+                                         :representation "plain"}}})}))))
 
 (defn tenant-name
   "